pages working, but not correctly

+++ src/main/java/com/lits/demo/repository/UserDataRepository.java
@@ -3,7 +3,7 @@ package com.lits.demo.repository;
 import com.lits.demo.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDataRepository extends JpaRepository <User, Integer> {
+public interface UserDataRepository extends JpaRepository <User, Long> {
 
     User findOneByUsername (String username);
 }
